pretty print records in updatable table records
