+++ package.json
@@ -35,6 +35,7 @@
     "@types/node": "^8.0.0",
     "@types/pify": "^0.0.28",
     "@types/sax": "^1.0.0",
+    "@types/sprintf-js": "^0.0.27",
     "@types/uuid": "^3.0.0",
     "@types/xml2js": "^0.0.33",
     "jest": "^20.0.4",
@@ -49,6 +50,7 @@
     "moment": "^2.18.1",
     "pify": "^3.0.0",
     "sax": "^1.2.2",
+    "sprintf-js": "^1.1.1",
     "uuid": "^3.1.0",
     "xml2js": "^0.4.17"
   }
